First skeleton of a python app for OpStack discovery.
Two classes for CPU and Mem (inspired by G-X attributes).
Plan is to use this script to talk to OpenStack APIs and fill
those classes (and more) with values.

For the OpenStack flavors, our intention is to parse the SCS
compliant flavor names to get more info than the standard
number of CPUs, mem, disk from the OpenStack API
(with openstack flavor show).

Signed-off-by: Kurt Garloff <kurt@garloff.de>