Change to output flavor list in format from MR34.
This aligns flavor output to the format suggested by
https://gitlab.com/gaia-x/technical-committee/service-characteristics/-/blob/76-brainstorming-on-infrastructure-service-offering-attribtues/single-point-of-truth/yaml/validation/flavor.yaml

Flavors are not a dict but a list with a (unique) name.
* The rationale for this could be discussed, but the format description
  may require a unique type for each dict and thus we're better off with
  the list.

Output diskSize and ramSize as gax-core:Measure data types.

We only output diskSize IF there is a disk. The cardinality allows for
this and the output looks shorter and cleaner, BUT it creates additional
burden on the consumer (to not silently assume diskSize to always be
there).

Comment out the spurious loop for in_count in range(len(flv_id)):

Signed-off-by: Kurt Garloff <kurt@garloff.de>