Output all services from service catalog.
We have implemented a generic mechanism to dump some
interesing info from the OpenStack service catalog
if we have support from the SDK.

osVolume, osCompute and osLoadBalancer are specializations for it.
- Compute does collect the flavor list
- Volume needs manual retrieval of extension list and AZs and
  does also have storage types
- LoadBalancer also has a flavor list

Also, we can output AZs from the network service and we outpu
regions for the cloud.

This is a fairly comprehensive list now, but of course work
remains:
- Probably some networking features to be detected and listed
- Flavor details (acc to existing and future SCS specs)
- Image data (incl SCS metadata info)

Signed-off-by: Kurt Garloff <kurt@garloff.de>