Handle multi-region catalogs. Handle conn errors.
We may have several regions listed in the service catalogue.
We need to handle the service list PER region. So sort the
public endpoints by region. Adjust output format to have
region as a high-level structural element below openstackcloud.

Some special treatment is needed to get version list and endpoints
specific to the region.

Also, we need to better handling of errors:
- Only probe endpoints when they are listed as public EPs for the
  region.
- Set lower timeout (default 12s, tunable with -t).
- Detect and handle failed connection setup (no api_version).

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Signed-off-by: Anja Strunk <anja.strunk@cloudandheat.com>