We had declared compute as being handled.
This was a stupid typo: When we have special classes for handling specific
OpenStack services, these would be remembered (with all their aliases) to
avoid being handled and listed twice. Due to a typo, we always recorded
the compute service as being handled already (instead of the one that we
actually handled specially, which include Compute, Volume, Network,
Loadbalancer).

This resulted in compute not being displayed if it was not the first out
of those four special classes in the service catalog.

Signed-off-by: Kurt Garloff <kurt@garloff.de>